+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.1.1 (Next)
 
+* Added support for converting and displaying measurements of activity distance, elevation, speed and pace - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.1.0 (2018/11/23)